Restructuring modules, ripping out Generic
